requirements.txt
mysql-connector-python>=8.0.0
sqlglot>=23.0.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
Flask>=3.1.0
